Remove test for the native database tracker
Ideally, finalizers added in the `sqlite3` package should replace this
functionality.  Currently, it seems like finalizers are not triggered by
hot restarts which is unfortunate, but we also can't test the feature
reliably since it now causes databases to be closed twice (once by
the test and once because sqlite3's finalizer runs).