Make explicit `schema generate` command for migration data tests
These flags are mentioned in the prose of this section, but it's
easy to miss them (or at least, I missed them) there and rely on
the `schema generate` command given in the previous section.
That one generates the expected files, but with this section's code
it leads to an error:

    Unsupported operation: TableInfo.map in schema verification code

because the generated code has a stub implementation of `map`.

Make it easier to spot the need for these flags by providing an
explicit complete command using them.