fix: Infer `json_extract` type from context
Currently, inference of all `json_extract` expressions results in an unknown type for inference. However,
there are cases where the type can be known/coerced, for example with `INSERT INTO SELECT` statements when
the column types are inferred from the written table.

In those cases, we use the expectation type to resolve `json_extract` expressions.